+++ pypsa/clustering/__init__.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any
 import pandas as pd
 
 from pypsa.clustering import spatial, temporal
-from pypsa.utils import deprecated_common_kwargs
+from pypsa.common import deprecated_common_kwargs
 
 if TYPE_CHECKING:
     from pypsa import Network
