+++ pypsa/components/abstract.py
@@ -1,7 +1,9 @@
 """
 Abstract components module.
 
-Contains classes and logic relevant to all component types in PyPSA.
+Contains classes and properties relevant to all component types in PyPSA. Also imports
+logic from other modules:
+- components.types
 """
 
 from __future__ import annotations
@@ -13,14 +15,14 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
 
 import geopandas as gpd
-import numpy as np
 import pandas as pd
 from pyproj import CRS
 
+from pypsa.common import equals
+from pypsa.components.descriptors import get_active_assets
 from pypsa.constants import DEFAULT_EPSG, DEFAULT_TIMESTAMP
 from pypsa.definitions.components import ComponentType
 from pypsa.definitions.structures import Dict
-from pypsa.utils import equals
 
 logger = logging.getLogger(__name__)
 
@@ -75,6 +77,12 @@ class Components(ComponentsData, ABC):
         It is not recommended to use this class outside of PyPSA.
 
     """
+
+    # Methods
+    # -------
+
+    # from pypsa.components.descriptors
+    get_active_assets = get_active_assets
 
     def __init__(
         self,
@@ -557,47 +565,6 @@ class Components(ComponentsData, ABC):
 
         """
         return self.dynamic
-
-    def get_active_assets(
-        self,
-        investment_period: int | str | Sequence | None = None,
-    ) -> pd.Series:
-        """
-        Get active components mask of componen type in investment period(s).
-
-        A component is considered active when:
-        - it's active attribute is True
-        - it's build year + lifetime is smaller than the investment period (if given)
-
-        Parameters
-        ----------
-        investment_period : int, str, Sequence
-            Investment period(s) to check for active within build year and lifetime. If
-            none only the active attribute is considered and build year and lifetime are
-            ignored. If multiple periods are given the mask is True if component is
-            active in any of the given periods.
-
-        Returns
-        -------
-        pd.Series
-            Boolean mask for active components
-
-        """
-        if investment_period is None:
-            return self.static.active
-        if not {"build_year", "lifetime"}.issubset(self.static):
-            return self.static.active
-
-        # Logical OR of active assets in all investment periods and
-        # logical AND with active attribute
-        active = {}
-        for period in np.atleast_1d(investment_period):
-            if period not in self.n_save.investment_periods:
-                raise ValueError("Investment period not in `n.investment_periods`")
-            active[period] = self.static.eval(
-                "build_year <= @period < build_year + lifetime"
-            )
-        return pd.DataFrame(active).any(axis=1) & self.static.active
 
 
 class SubNetworkComponents:
