+++ pypsa/consistency.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING, Any
 import numpy as np
 import pandas as pd
 
-from pypsa.utils import deprecated_common_kwargs
+from pypsa.common import deprecated_common_kwargs
 
 if TYPE_CHECKING:
     from pypsa import Network
