+++ test/test_docs.py
@@ -12,6 +12,7 @@ import pypsa
 modules = [
     importlib.import_module(name)
     for _, name, _ in pkgutil.walk_packages(pypsa.__path__, pypsa.__name__ + ".")
+    if name not in ["pypsa.utils", "pypsa.components.utils"]
 ]
